Refactor IOButton to be called under HAL
